Move sortLeaderboard into toDto mapping
