+++ src/main/java/com/rocketden/main/service/SubmitService.java
@@ -14,7 +14,6 @@ import com.rocketden.main.model.problem.Problem;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import java.util.Comparator;
 import java.util.List;
 
 /**
@@ -55,38 +54,11 @@ public class SubmitService {
 
         // Sort list of players by who is winning
         GameDto gameDto = GameMapper.toDto(game);
-        sortLeaderboard(gameDto.getPlayers());
 
         // Send socket update with latest leaderboard info
         // TODO: test socket in GameSocketTests when merged
         socketService.sendSocketUpdate(gameDto);
 
         return GameMapper.submissionToDto(submission);
-    }
-
-    // Sort by numCorrect followed by startTime
-    public void sortLeaderboard(List<PlayerDto> players) {
-        players.sort((player1, player2) -> {
-            List<SubmissionDto> submissions1 = player1.getSubmissions();
-            List<SubmissionDto> submissions2 = player2.getSubmissions();
-
-            // Players who haven't submitted yet are sorted last
-            if (submissions1.isEmpty()) {
-                return 1;
-            } else if (submissions2.isEmpty()) {
-                return -1;
-            }
-
-            SubmissionDto sub1 = submissions1.get(submissions1.size() - 1);
-            SubmissionDto sub2 = submissions2.get(submissions2.size() - 1);
-
-            // If both have the same numCorrect, whoever submits earlier is first
-            if (sub1.getNumCorrect().equals(sub2.getNumCorrect())) {
-                return sub1.getStartTime().compareTo(sub2.getStartTime());
-            }
-
-            // Whoever has higher numCorrect is first
-            return sub2.getNumCorrect() - sub1.getNumCorrect();
-        });
     }
 }
