+++ src/test/java/com/rocketden/main/mapper/GameMapperTests.java
@@ -26,6 +26,7 @@ import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertNull;
 
+import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -44,6 +45,15 @@ public class GameMapperTests {
 
     private static final String INPUT = "[1, 8, 2]";
     private static final String OUTPUT = "[1, 2, 8]";
+
+    // Helper method to add a dummy submission to a PlayerDto object
+    private void addSubmissionHelper(PlayerDto playerDto, int numCorrect) {
+        SubmissionDto submissionDto = new SubmissionDto();
+        submissionDto.setNumCorrect(numCorrect);
+        submissionDto.setStartTime(LocalDateTime.now());
+
+        playerDto.getSubmissions().add(submissionDto);
+    }
 
     @Test
     public void fromRoom() {
@@ -152,5 +162,40 @@ public class GameMapperTests {
         assertEquals(submission.getPlayerCode().getLanguage(), submissionDto.getLanguage());
         assertEquals(submission.getNumCorrect(), submissionDto.getNumCorrect());
         assertEquals(submission.getStartTime(), submissionDto.getStartTime());
+    }
+
+    @Test
+    public void sortLeaderboardSuccess() {
+        List<PlayerDto> players = new ArrayList<>();
+
+        PlayerDto player1 = new PlayerDto();
+        addSubmissionHelper(player1, 0);
+
+        PlayerDto player2 = new PlayerDto();
+        addSubmissionHelper(player2, 0);
+        addSubmissionHelper(player2, 3);
+
+        PlayerDto player3 = new PlayerDto();
+        addSubmissionHelper(player3, 3);
+
+        PlayerDto player4 = new PlayerDto();
+        addSubmissionHelper(player4, 5);
+
+        PlayerDto player5 = new PlayerDto();
+
+        players.add(player1);
+        players.add(player2);
+        players.add(player3);
+        players.add(player4);
+        players.add(player5);
+
+        // Player order should be: [4, 2, 3, 1, 5]
+        GameMapper.sortLeaderboard(players);
+
+        assertEquals(player4, players.get(0));
+        assertEquals(player2, players.get(1));
+        assertEquals(player3, players.get(2));
+        assertEquals(player1, players.get(3));
+        assertEquals(player5, players.get(4));
     }
 }
